[refactor] restructured code to make it modular
