+++ public/index.html
@@ -31,7 +31,7 @@
         <span class="close-btn">x</span>
         <p id="errorText"></p>
     </div>    
-    <script src="app.js"></script>
+    <script type="module" src="/scripts/app.js"></script>
     <script src="handleSpotifyUserAuthentication.js"></script>
 </body>
