+++ public/scripts/modules/restoreUserInput.js
@@ -1,0 +1,7 @@
+export const restoreBookName = () => {
+    const bookName = localStorage.getItem('bookName');
+    if (bookName) {
+        document.getElementById('bookName').value = bookName;
+        localStorage.removeItem('bookName');
+    }
+};
